+++ src/main/java/Main.java
@@ -1,0 +1,10 @@
+package main.java;
+
+public class Main {
+    public static void main(String[] args) throws Exception {
+        CiServer srv = new CiServer(8080);
+        srv.join();
+
+
+    }
+}
